refactor(content): move flows into content/flows.ts
Flows (daily notes) were already self-contained: FlowSchema, the
YYYY/MM/DD tree walk, and date-derived slugs only needed io + text
metrics. getFlowTags moves with them — it is a flows aggregate, not a
cross-content one. 12 consumers update; the static-params unit test
stubs the new module. feed-utils' same-directory './markdown' import was
missed by the migration script's path filter — fixed here and the script
pattern corrected for the remaining phases.
