+++ src/app/flows/[year]/[month]/page.tsx
@@ -1,4 +1,4 @@
-import { getAllFlows, getFlowsByMonth, getFlowTags } from '@/lib/markdown';
+import { getAllFlows, getFlowsByMonth, getFlowTags } from '@/lib/content/flows';
 import { siteConfig } from '../../../../../site.config';
 import { Metadata } from 'next';
 import { notFound } from 'next/navigation';
