+++ src/lib/feed-utils.ts
@@ -4,7 +4,8 @@ import remarkGfm from 'remark-gfm';
 import remarkRehype from 'remark-rehype';
 import rehypeRaw from 'rehype-raw';
 import rehypeStringify from 'rehype-stringify';
-import { getAllPosts, getAllFlows } from './markdown';
+import { getAllPosts } from './markdown';
+import { getAllFlows } from './content/flows';
 import { siteConfig } from '../../site.config';
 import { getPostUrl, getFlowUrl } from './urls';
 import { resolveLocale } from './i18n';
